hns3 udma: support WRITE With Notify
driver inclusion
category: feature
bugzilla: https://gitee.com/openeuler/kernel/issues/I85R2F
CVE: NA

-----------------------------------------------------------

The process of WRITE With Notify is as follows:
1. The receiver creates a JFC of the notify type to write
parameters such as the notify address to the CQC.
2. A user uses the post_JFS/jetty_wr flow to deliver a WR.
After the driver identifies that the operation code of urma is
URMA_OPC_WRITE_NOTIFY, Parses and fills the notify information
(including the address offset and notify data) transferred by
the user to the WQE for hardware to read.
3. After the hardware reads the data, the notify data is
written to the specified address. The formula is as follows:
Specified address = Notify address specified by the JFC +
Address offset transferred by the I/O.

Signed-off-by: Lin He <helin52@huawei.com>